Implement peer.server.Server::select()
This method accepts sockets to additionally select() on and call the
associated action whenever data comes available. The callback is passed
the server instance and the socket. This can be used to shut down the
server, for instance, or implement server push, as seen in this chat
application example: https://gist.github.com/thekid/275f0cfcf90a09d79c65c86c61b60b25