Implement asynchronous server
It supports protocols where handleData() returns a generator function to yield
back control to the server, e.g. during lengthy up- or download operations.

For backwards compatibility, this is implemented in a new server class called
peer.server.AsyncServer. However, this will become the default with the next
major release